+++ lib/driver/_cdio_stream.c
@@ -149,13 +149,13 @@ _cdio_stream_open_if_necessary(CdioDataSource_t *p_obj)
 ssize_t
 cdio_stream_read(CdioDataSource_t* p_obj, void *ptr, size_t size, size_t nmemb)
 {
-  long read_bytes;
+  ssize_t read_bytes;
 
   if (!p_obj) return 0;
   if (!_cdio_stream_open_if_necessary(p_obj)) return 0;
 
   read_bytes = (p_obj->op.read)(p_obj->user_data, ptr, size*nmemb);
-  p_obj->position += read_bytes;
+  p_obj->position += (off_t)read_bytes;
 
   return read_bytes;
 }
